--- src/SME.SGP.Aplicacao/Commands/Fechamento/SalvarFechamentoTurma/SalvarFechamentoTurmaCommand.cs
@@ -1,16 +1,0 @@
-﻿using MediatR;
-using SME.SGP.Dominio;
-using SME.SGP.Infra;
-
-namespace SME.SGP.Aplicacao
-{
-    public class SalvarFechamentoTurmaCommand : IRequest<long>
-    {
-        public SalvarFechamentoTurmaCommand(FechamentoTurma fechamentoTurma)
-        {
-            FechamentoTurma = fechamentoTurma;
-        }
-
-        public FechamentoTurma FechamentoTurma { get; }
-    }
-}
